// types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      applications: {
        Row: {
          application_id: number
          approach: string | null
          availability: string
          created_at: string
          cv: string | null
          email: string
          evil_score: number | null
          expertise: string | null
          first_name: string
          hierarchy: string | null
          job_id: number | null
          last_name: string
          risk: string | null
          salary: number
          status: string | null
          user_id: string | null
        }
        Insert: {
          application_id?: number
          approach?: string | null
          availability: string
          created_at?: string
          cv?: string | null
          email: string
          evil_score?: number | null
          expertise?: string | null
          first_name: string
          hierarchy?: string | null
          job_id?: number | null
          last_name: string
          risk?: string | null
          salary: number
          status?: string | null
          user_id?: string | null
        }
        Update: {
          application_id?: number
          approach?: string | null
          availability?: string
          created_at?: string
          cv?: string | null
          email?: string
          evil_score?: number | null
          expertise?: string | null
          first_name?: string
          hierarchy?: string | null
          job_id?: number | null
          last_name?: string
          risk?: string | null
          salary?: number
          status?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "applications_job_id_fkey"
            columns: ["job_id"]
            isOneToOne: false
            referencedRelation: "jobs"
            referencedColumns: ["job_id"]
          },
        ]
      }
      jobs: {
        Row: {
          approach: string
          created_at: string
          department: string
          description: string
          expertise: string
          hierarchy: string
          job_id: number
          risk: string
          skills: string[]
          tasks: string[]
          title: string
        }
        Insert: {
          approach: string
          created_at?: string
          department: string
          description: string
          expertise: string
          hierarchy: string
          job_id?: number
          risk: string
          skills: string[]
          tasks: string[]
          title: string
        }
        Update: {
          approach?: string
          created_at?: string
          department?: string
          description?: string
          expertise?: string
          hierarchy?: string
          job_id?: number
          risk?: string
          skills?: string[]
          tasks?: string[]
          title?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never